rename Analytics module that was causing problems with filename case and redisgn playback control icons in player ui
